Create README.md
The video game, Fallout 3, has a computer hacking mini game in which the user tries to guess the work to hack into the computer.

This game requires the player to correctly guess a password from a list of same length words. The player has only 4 guesses and on each incorrect guess the computer will indicate how many letter positions are correct.

For example, if the password is MIND and the player guesses MEND, the game will indicate that 3 out of 4 positions are correct (M_ND). If the password is COMPUTE and the player guesses PLAYFUL, the game will report 0/7. While some of the letters match, they're in the wrong position.

Ask the player for a difficulty (very easy, easy, average, hard, very hard), then present the player with 3 to 15 words of the same length. The player then has 4 guesses, and on each incorrect guess indicate the number of correct positions.

Example:

FALLOUT INDUSTRIES (TM) TERMALINK PROTOCOL
ENTER DIFFICULTY(1 - 5): 2

OPENING TERMINAL. ENTER CREDENTIALS TO PROCEED...


0x01980998   STERES   0x019809B4   CHROME
0x019809D0   DRYADS   0x019809EC   YESSED
0x01980A08   SHUNTS   0x01980A24   OXALIC
0x01980A40   WIFELY   0x01980A5C   TANGLE
0x01980A78   YABBER   0x01980A94   BROLLY

4 ATTEMPT(S) LEFT: steres
0/6 correct

ENTRY DENIED


3 ATTEMPT(S) LEFT: chrome
1/6 correct

ENTRY DENIED


2 ATTEMPT(S) LEFT: shunts
0/6 correct

CONSOLE LOCKDOWN IMMINENT!!!

ENTRY DENIED


1 ATTEMPT(S) LEFT: wifley
0/6 correct

ENTRY DENIED


The correct password was: TANGLE

LOGIN FAIL. CLOSING PROGRAM...

Press any key to continue . . .